requirements.txt
PyQt6
opencv-python
numpy
mss
pyautogui
Pillow
easyocr